Simple calculator program in Python
A very basic calculator function that allows users to perform addition, subtraction, multiplication, and division, running in a loop until the user chooses to exit. The user is prompted to select an operation and enter 2 numeric values as operands. Invalid numerical inputs are handled and division by zero is prevented. At the end of each calculation, the user is asked if they would like to perform another calculation and the program either loops or exits depending on their selection. If the user enters an invalid input at the continuation prompt they will be repeatedly asked until a valid input is entered.